+++ clang-tools-extra/clangd/Config.h
@@ -114,7 +114,7 @@ struct Config {
       /// IncludeCleaner will not diagnose usages of these headers matched by
       /// these regexes.
       std::vector<std::function<bool(llvm::StringRef)>> IgnoreHeader;
-      bool AnalyzeSystemHeaders = false;
+      bool AnalyzeAngledIncludes = false;
     } Includes;
   } Diagnostics;