+++ clang-tools-extra/clangd/ConfigYAML.cpp
@@ -169,9 +169,9 @@ private:
       if (auto Values = scalarValues(N))
         F.IgnoreHeader = std::move(*Values);
     });
-    Dict.handle("AnalyzeSystemHeaders", [&](Node &N) {
-      if (auto Value = boolValue(N, "AnalyzeSystemHeaders"))
-        F.AnalyzeSystemHeaders = *Value;
+    Dict.handle("AnalyzeAngledIncludes", [&](Node &N) {
+      if (auto Value = boolValue(N, "AnalyzeAngledIncludes"))
+        F.AnalyzeAngledIncludes = *Value;
     });
     Dict.parse(N);
   }