+++ clang-tools-extra/clangd/unittests/ConfigCompileTests.cpp
@@ -279,10 +279,10 @@ TEST_F(ConfigCompileTests, DiagnosticsIncludeCleaner) {
   EXPECT_FALSE(HeaderFilter("bar.h"));
 
   Frag = {};
-  EXPECT_FALSE(Conf.Diagnostics.Includes.AnalyzeSystemHeaders);
-  Frag.Diagnostics.Includes.AnalyzeSystemHeaders = true;
+  EXPECT_FALSE(Conf.Diagnostics.Includes.AnalyzeAngledIncludes);
+  Frag.Diagnostics.Includes.AnalyzeAngledIncludes = true;
   EXPECT_TRUE(compileAndApply());
-  EXPECT_TRUE(Conf.Diagnostics.Includes.AnalyzeSystemHeaders);
+  EXPECT_TRUE(Conf.Diagnostics.Includes.AnalyzeAngledIncludes);
 }
 
 TEST_F(ConfigCompileTests, DiagnosticSuppression) {